requirements.txt
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0
scikit-learn>=1.3.0
xgboost>=2.0.0
prophet>=1.1.4
torch>=2.0.0
requests>=2.31.0
pyyaml>=6.0
jupyter>=1.0.0
optuna>=3.3.0
shap>=0.42.0
folium>=0.14.0
tqdm>=4.65.0
openpyxl>=3.1.0
